Storing values in double-width floating point numbers, to increase precision with small numbers

+++ src/main.c
@@ -102,7 +102,7 @@ int main(int argc, const char** argv) {
     } else if(strcmp(p.solver, "dp") == 0) {
         TBKPDPStats stats;
 
-        const float sol = tbkp_dp_solve(instance, &stats, &p);
+        const double sol = tbkp_dp_solve(instance, &stats, &p);
         printf("Dynamic programming. Solution: %.2f\n", sol);
         tbkp_dp_stats_to_file(&stats, &p);
     } else {
